+++ evalscope/tools/rewrite_eval_results.py
@@ -4,7 +4,7 @@ import time
 
 from evalscope.models.custom import CustomModel
 from evalscope.run import run_task
-from evalscope.utils import yaml_to_dict
+from evalscope.utils.io_utils import yaml_to_dict
 from evalscope.utils.logger import get_logger
 
 logger = get_logger()
